Moved edit form state to component
There were performance problems with storing the changeset within the redux
store. Most of the application was having to re-render itself because we were
updating fields on the EditForm.

The solution was to move storing the state to the component and keep it out of
redux store. This way the only thing that needs to re-render is the form itself.